csv_types: Add a Source class to define custom details parsers
We could try to do something smart and use ML/AI to make things work for any
random source type, but without enough data it would be pretty hard to build
something that works well. It may be best to allow people to write and/or
contribute parser for their own source types and may be eventually build a more
global/automated way to do this.