model: Allow only category per transaction
Multiple categories make it harder to categorize where each rupee is going
during a particular time duration. It also makes it harder to categorize
transactions automatically when each transaction can have multiple
categories. This commit makes it easier to visualize categories data and
automate categorization of transactions.

In a future commit, it is possible to add tags to allow more granular tagging
to allow for more visualizations and better searching of transactions.