app: Fix inplace set on a copy via chained assignment
Fixes the following warnings from Pandas:

FutureWarning: A value is trying to be set on a copy of a DataFrame or
Series through chained assignment using an inplace method. The behavior
will change in pandas 3.0. This inplace method will never work because
the intermediate object on which we are setting values always behaves as
a copy.

For example, when doing 'df[col].method(value, inplace=True)', try using
'df.method({col: value}, inplace=True)' or df[col] =
df[col].method(value) instead, to perform the operation inplace on the
original object.