parse: Use filename + row-number in the unique sha
Axis bank has conveniently removed the transaction number or any other
unique identifier from the transaction details column. This makes the
parsing and identifying unique rows impossible. To workaround that we
use the filename and the row number to uniquely identify transctions.
Since, we have changed the scraping to only get monthly data in each
file, this should work alright.